RedShadow/requirements.txt
requests
pycryptodome
sqlite3
nmap
bs4